+++ src/vtt/__init__.py
@@ -1,0 +1,1 @@
+"""Unified VoiceToText CLI package."""